+++ release-raccoon-app/src/main/java/com/raccoon/dto/mapping/ArtistMapper.java
@@ -4,9 +4,11 @@ import com.raccoon.dto.ArtistDto;
 import com.raccoon.entity.Artist;
 import com.raccoon.search.dto.SearchResultArtistDto;
 
+import org.mapstruct.AfterMapping;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingConstants;
+import org.mapstruct.MappingTarget;
 
 @Mapper(componentModel = MappingConstants.ComponentModel.CDI, uses = ArtistMapperResolver.class)
 public interface ArtistMapper {
@@ -17,5 +19,15 @@ public interface ArtistMapper {
 
     @Mapping(target = "id", ignore = true)
     Artist fromDto(SearchResultArtistDto dto);
+
+    @AfterMapping
+    default void addFollowerCount(@MappingTarget ArtistDto artistDto, Artist artist) {
+        // Follower count will be set by the service layer to avoid lazy loading issues
+    }
+
+    @AfterMapping
+    default void addFollowerCount(@MappingTarget SearchResultArtistDto artistDto, Artist artist) {
+        // Follower count will be set by the service layer to avoid lazy loading issues
+    }
 
 }